% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{as_numberwang}
\alias{as_numberwang}
\alias{as_numberwang.default}
\alias{as_numberwang.character}
\title{Convert to numberwang}
\usage{
as_numberwang(x, ...)

\method{as_numberwang}{default}(x, ...)

\method{as_numberwang}{character}(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{ignored}
}
\description{
Convert to numberwang
}
