% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{fixedpt}
\alias{fixedpt}
\title{Create a vector of \code{fixedpt} values from numeric values}
\usage{
fixedpt(x = double(), decimals = 3)
}
\arguments{
\item{x}{vector of doubles}

\item{decimals}{number of decimal places required in representation}
}
\value{
fixedpt vector
}
\description{
Create a vector of \code{fixedpt} values from numeric values
}
