% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{format.typ_fixedpt}
\alias{format.typ_fixedpt}
\title{Format a \code{fixedpt} value}
\usage{
\method{format}{typ_fixedpt}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Format a \code{fixedpt} value
}
