% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{format.typ_numberwang}
\alias{format.typ_numberwang}
\title{Format a \code{numberwang} value}
\usage{
\method{format}{typ_numberwang}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Format a \code{numberwang} value
}
