% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{is_numberwang}
\alias{is_numberwang}
\title{Test if object is \code{numberwang}}
\usage{
is_numberwang(x)
}
\arguments{
\item{x}{object}
}
\value{
logical
}
\description{
Test if object is \code{numberwang}
}
