% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{new_fixedpt}
\alias{new_fixedpt}
\title{New fixed point}
\usage{
new_fixedpt(x = double(), decimals = 3)
}
\arguments{
\item{x}{vector of doubles}

\item{decimals}{number of decimal places required in representation}
}
\description{
New fixed point
}
