% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{new_numberwang}
\alias{new_numberwang}
\title{New word}
\usage{
new_numberwang(x = double(), decimals = 3)
}
\arguments{
\item{x}{vector of doubles}

\item{decimals}{how many decimals to show in output. default: 3}
}
\description{
New word
}
