% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{numberwang}
\alias{numberwang}
\title{Create a vector of \code{numberwang} values from double values}
\usage{
numberwang(x = double(), decimals = 3)
}
\arguments{
\item{x}{vector of doubles}

\item{decimals}{how many decimals to show in output. default: 3}
}
\value{
numberwang vector
}
\description{
Create a vector of \code{numberwang} values from double values
}
