% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{vec_arith.typ_fixedpt}
\alias{vec_arith.typ_fixedpt}
\title{Arithmetic ops on \code{fixedpt}}
\usage{
\method{vec_arith}{typ_fixedpt}(op, x, y, ...)
}
\arguments{
\item{op}{function}

\item{x, y}{lhs, rhs}

\item{...}{ignored}
}
\description{
Note: results will have as many decomals as the maximum of the two arguments,
so there's always the posibility of overflow.
}
