% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{vec_arith.typ_numberwang}
\alias{vec_arith.typ_numberwang}
\title{Arithmetic ops}
\usage{
\method{vec_arith}{typ_numberwang}(op, x, y, ...)
}
\arguments{
\item{op}{function}

\item{x, y}{lhs, rhs}

\item{...}{ignored}
}
\description{
Arithmetic ops
}
