% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint16.R
\name{vec_arith.typ_uint16}
\alias{vec_arith.typ_uint16}
\title{Arithmetic ops}
\usage{
\method{vec_arith}{typ_uint16}(op, x, y, ...)
}
\arguments{
\item{op}{function}

\item{x, y}{lhs, rhs}

\item{...}{ignored}
}
\description{
Arithmetic ops
}
