% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{vec_cast.typ_fixedpt.typ_fixedpt}
\alias{vec_cast.typ_fixedpt.typ_fixedpt}
\alias{vec_cast.typ_fixedpt.double}
\alias{vec_cast.double.typ_fixedpt}
\alias{vec_cast.character.typ_fixedpt}
\title{Cast between \code{fixedpt} and \code{double}}
\usage{
\method{vec_cast}{typ_fixedpt.typ_fixedpt}(x, to, ...)

\method{vec_cast}{typ_fixedpt.double}(x, to, ...)

\method{vec_cast}{double.typ_fixedpt}(x, to, ...)

\method{vec_cast}{character.typ_fixedpt}(x, to, ...)
}
\arguments{
\item{x}{object}

\item{to}{conversion target type}

\item{...}{ignored}
}
\description{
Cast between \code{fixedpt} and \code{double}
}
