% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{vec_cast.typ_numberwang.typ_numberwang}
\alias{vec_cast.typ_numberwang.typ_numberwang}
\alias{vec_cast.typ_numberwang.double}
\alias{vec_cast.double.typ_numberwang}
\alias{vec_cast.typ_numberwang.integer}
\alias{vec_cast.integer.typ_numberwang}
\alias{vec_cast.character.typ_numberwang}
\title{Casting}
\usage{
\method{vec_cast}{typ_numberwang.typ_numberwang}(x, to, ...)

\method{vec_cast}{typ_numberwang.double}(x, to, ...)

\method{vec_cast}{double.typ_numberwang}(x, to, ...)

\method{vec_cast}{typ_numberwang.integer}(x, to, ...)

\method{vec_cast}{integer.typ_numberwang}(x, to, ...)

\method{vec_cast}{character.typ_numberwang}(x, to, ...)
}
\arguments{
\item{x}{object}

\item{to}{destination prototype}

\item{...}{other arguments}
}
\description{
Casting
}
