% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint16.R
\name{vec_cast.typ_uint16.typ_uint16}
\alias{vec_cast.typ_uint16.typ_uint16}
\alias{vec_cast.typ_uint16.integer}
\alias{vec_cast.integer.typ_uint16}
\alias{vec_cast.typ_uint16.double}
\alias{vec_cast.double.typ_uint16}
\alias{vec_cast.character.typ_uint16}
\title{Casting}
\usage{
\method{vec_cast}{typ_uint16.typ_uint16}(x, to, ...)

\method{vec_cast}{typ_uint16.integer}(x, to, ...)

\method{vec_cast}{integer.typ_uint16}(x, to, ...)

\method{vec_cast}{typ_uint16.double}(x, to, ...)

\method{vec_cast}{double.typ_uint16}(x, to, ...)

\method{vec_cast}{character.typ_uint16}(x, to, ...)
}
\arguments{
\item{x}{object}

\item{to}{destination prototype}

\item{...}{other arguments}
}
\description{
Casting
}
