% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{vec_math.typ_fixedpt}
\alias{vec_math.typ_fixedpt}
\title{Math function calls on \code{fixedpt}}
\usage{
\method{vec_math}{typ_fixedpt}(.fn, .x, ...)
}
\arguments{
\item{.fn}{function name as string}

\item{.x}{fixedpt object}

\item{...}{other arguments to \code{.fn}}
}
\description{
Math function calls on \code{fixedpt}
}
