% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{vec_math.typ_numberwang}
\alias{vec_math.typ_numberwang}
\title{Math function calls}
\usage{
\method{vec_math}{typ_numberwang}(.fn, .x, ...)
}
\arguments{
\item{.fn}{function name as string}

\item{.x}{object}

\item{...}{other arguments to \code{.fn}}
}
\description{
Math function calls
}
