% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint16.R
\name{vec_math.typ_uint16}
\alias{vec_math.typ_uint16}
\title{Math function calls}
\usage{
\method{vec_math}{typ_uint16}(.fn, .x, ...)
}
\arguments{
\item{.fn}{function name as string}

\item{.x}{object}

\item{...}{other arguments to \code{.fn}}
}
\description{
Math function calls
}
