% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{vec_ptype2.typ_fixedpt.typ_fixedpt}
\alias{vec_ptype2.typ_fixedpt.typ_fixedpt}
\alias{vec_ptype2.typ_fixedpt.double}
\alias{vec_ptype2.double.typ_fixedpt}
\title{Coercion between \code{fixedpt} and \code{double}}
\usage{
\method{vec_ptype2}{typ_fixedpt.typ_fixedpt}(x, y, ...)

\method{vec_ptype2}{typ_fixedpt.double}(x, y, ...)

\method{vec_ptype2}{double.typ_fixedpt}(x, y, ...)
}
\arguments{
\item{x, y}{objects}

\item{...}{ignored}
}
\description{
Coercion between \code{fixedpt} and \code{double}
}
