% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberwang.R
\name{vec_ptype2.typ_numberwang.typ_numberwang}
\alias{vec_ptype2.typ_numberwang.typ_numberwang}
\alias{vec_ptype2.typ_numberwang.double}
\alias{vec_ptype2.double.typ_numberwang}
\alias{vec_ptype2.typ_numberwang.integer}
\alias{vec_ptype2.integer.typ_numberwang}
\title{Coercion}
\usage{
\method{vec_ptype2}{typ_numberwang.typ_numberwang}(x, y, ...)

\method{vec_ptype2}{typ_numberwang.double}(x, y, ...)

\method{vec_ptype2}{double.typ_numberwang}(x, y, ...)

\method{vec_ptype2}{typ_numberwang.integer}(x, y, ...)

\method{vec_ptype2}{integer.typ_numberwang}(x, y, ...)
}
\arguments{
\item{x, y}{objects}

\item{...}{other arguments}
}
\description{
Coercion
}
