% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint16.R
\name{vec_ptype2.typ_uint16.typ_uint16}
\alias{vec_ptype2.typ_uint16.typ_uint16}
\alias{vec_ptype2.typ_uint16.integer}
\alias{vec_ptype2.integer.typ_uint16}
\alias{vec_ptype2.typ_uint16.double}
\alias{vec_ptype2.double.typ_uint16}
\title{Coercion}
\usage{
\method{vec_ptype2}{typ_uint16.typ_uint16}(x, y, ...)

\method{vec_ptype2}{typ_uint16.integer}(x, y, ...)

\method{vec_ptype2}{integer.typ_uint16}(x, y, ...)

\method{vec_ptype2}{typ_uint16.double}(x, y, ...)

\method{vec_ptype2}{double.typ_uint16}(x, y, ...)
}
\arguments{
\item{x, y}{objects}

\item{...}{other arguments}
}
\description{
Coercion
}
