% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedpt.R
\name{vec_ptype_abbr.typ_fixedpt}
\alias{vec_ptype_abbr.typ_fixedpt}
\title{Abbreviation to use in some outputs e.g. \code{tibble}}
\usage{
\method{vec_ptype_abbr}{typ_fixedpt}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Abbreviation to use in some outputs e.g. \code{tibble}
}
